Not logging the runtime of previous input discovery so that the critical path can report more actions.
Since we are utilizing the start time of an input discovery of a restarted action, there will be a huge gap between then and when the action actually completes. With the current implementation of the critical path, this will block other actions that are dependencies of the original action from being reported on the critical path.

RELNOTES: None.
PiperOrigin-RevId: 292402157