Automated rollback of commit e0d325b44b9941931aa3c0d94fb2e446d60ba98c.
*** Reason for rollback ***

This caused a P1 breakage: local code changes are not reflected in incremental runs.

*** Original change description ***

"Stream" action cache digest computations.

Now that ActionAnalysisMetadata.getInputs() returns a NestedSet, we don't need allocate an intermediate (path -> metadata) map during action digest computation. Encapsulate the logic for this in a OrderIndependentHasher class. This class also caches fingerprints and digest arrays to avoid allocations.

Update the size constant in CompactPersistentActionCacheTest.testToStringIsntToBig because the digests are all 32 byte arrays instead of 1 byte arrays.

Additionally, fix ActionCache.Entry so it's stateless and immutable by making a builder class for it. Delete CompactPersistentActionCacheTest.testEntryToStringIsIdempotent because the property it wants to proven by the type system.

Closes #10537.

***

RELNOTES: None
PiperOrigin-RevId: 292894801