bazel syntax: reduce log spam from tests
ParserTest and ValidationTest no longer print their
expected errors to the log. The root problem is in
EventCollectionApparatus, which prints all errors,
but it is so widely used that changing it is a bigger
project. This change simply removes the two tests'
dependency on the apparatus, which was overkill.

Also:
- make tests final to prevent subclassing, where possible.
  (If you make it private it silently fails to run.)
- Simplify code to parse semantic options.
PiperOrigin-RevId: 293006744