Improve mechanism to distribute precompiled jars (e.g. for bootstrapping)
This change adds a new directory `derived/jars` to `bazel-distfile.{tar,zip}` that contains jars that we can't (or don't want to) build during bootstrapping (e.g. the Protobuf runtime).
Delaying the compilation of these jars until we create the distfile archive removes the need to check-in these precompiled jars.

This should allow us to replace some of the checked-in third-party repositories with `http_archive`.

Closes #10638.

PiperOrigin-RevId: 293169661