Fix getChildrenInterruptibly to actually propagate the InterruptedException.
In https://github.com/bazelbuild/bazel/commit/6b2e3e043a49f5ec28c4d03457a3e0ca50010b83, I mistakenly thought that handleInterruptedException=true meant that it would be propagated, but it's actually the opposite. So that this mistake is not made in vain, I've created an enum to make the behavior more explicit and added a test case.

RELNOTES: None.
PiperOrigin-RevId: 293179049