On OpenBSD, search the PATH to find the Bazel executable.
Searching the `PATH` is the only feasible way to find the executable on OpenBSD when `argv[0]` is not an absolute path or a relative path. This change resolves a TODO.

This change moves a preexisting `Which` function for searching the `PATH` out of `blaze_util_linux.cc` and into `blaze_util_posix.cc`, so that the Linux code and the BSD code can share this function.

In my testing on OpenBSD 6.6-current, a bootstrap build of Bazel succeeds and the resulting `bazel` binary can find itself on the `PATH`. (One caveat: For the bootstrap build to succeed, I had to manually apply the unrelated change in https://github.com/bazelbuild/bazel/pull/10639, since it's not merged yet.)

This change is part of the OpenBSD port in https://github.com/bazelbuild/bazel/issues/10250.

@jmmv FYI.

Closes #10691.

PiperOrigin-RevId: 293183553