Fix a deficiency in WorkerExecRootTest
The test should've simulated the existence of files in exec root
not the workspace dir. This meant the test did not check whether
old symlinks get removed (although it did correctly check the
creation of requested symlinks)

The change also adds an additional test to make sure we do not
delete files in the workspace directory.

RELNOTES: None.
PiperOrigin-RevId: 293330213