Following CL by lberki@, report all errors in analysis tests.
Before, loading phase errors and errors stemming from a non-existent top-level target were simply swallowed.

There is still a divergence between the behavior of SkyframeBuildView in tests and in production code and the theoretically best fix would be to avoid the long chain of *ForTesting methods, but in the meantime, having one such condition instead of two is a net win.

A few tests had to be fixed that erroneously passed due to the swallowed errors.

PiperOrigin-RevId: 293419810