Add support for bash completion of --config= names
This change supports the following command line arguments:
--bazelrc=<file>
--nosystem_rc
--noworkspace_rc
--nohome_rc

And supports finding bazelrc files in all the documented locations
as-per the docs at:
https://docs.bazel.build/versions/2.0.0/guide.html#bazelrc

It can handle command inheritance. So if you specify a "test" command
it knows to look at test, build, and common configs.

It also can deal with import and try-import commands, and can chase
those down to any nesting level.