bazel syntax: delete Callstack (used by memory allocation tracker)
The Callstack was a copy of the thread's stack saved in thread-local
storage. It was notified not only of every function call entry and
exit, but the same for every expression too. This made it not only
intrusive but incompatible with a compiled representation.

This change deletes Callstack and replaces it with an interface,
Debug.ThreadHook, that lets a single client be notified when a
thread pushes its first frame or pops its last.
The allocation tracker uses these notifications to maintain a mapping
of Java threads to StarlarkThreads (similar to CpuProfiler),
so that when the memory allocator hook is called, it can ascertain
which Starlark thread caused the allocation.

The tests have been reworked to call the interpreter instead of
peeking and poking into the callstack directly.

BUG=149023294
PiperOrigin-RevId: 293629363