Fix a memory leak in linux sandbox
`getcwd` result is allocated using `malloc` and should be deallocated with
`free`.

However `opt.working_dir` won't own it because that's an `std::string` and
the current `operator=` call will allocate a new buffer and copy the
content of `getcwd` result.

Hence, in the previous solution, the result of `getcwd` was just leaked at
the end of the function.

This commit adds the necessary `free`.

*edit*: I found that when running a C++ ASAN sanitizer during my build. I have a rule which setups the environment with `LD_PRELOAD and apparently bazel is calling the sandbox inside the build environment, hence it inherit the `LD_PRELOAD` setting.

Closes #10599.

PiperOrigin-RevId: 293835331