MockToolsConfig now enforces relative path args.
In MockToolsConfig, all rootDirectory.getRelative
calls except the one in the ctor (and in getPath)
are replaced with a call to getPath, because this
standardizes all path lookups.
The ctor's call is legitimate and remains, because
it converts an absolute path string to a Path
object and there's no other way to do that.

Also, MockToolsConfig.getPath now validates that
its argument is a relative path, because that's
what the function expects (according to the
variable name) and because that's how we can
support real (not in-memory) filesystems.

Finally, fix call sites that passed absolute paths
(e.g. create("/local_config_xcode/WORKSPACE")) to
pass a relative path instead, again because that's
how we can use a real filesystem. (All paths will
be made relative to the FileSystem's root
directory.)

RELNOTES: none
PiperOrigin-RevId: 294155732