Remove git notes usage from release process
Previously we store releae name, candidate number, cherry-picks and
release notes info in git repo using `git notes`.  To remove this
complexity, we remove all git notes usage in this change.

Instead of using git notes, we parse release name and candidate number
from branch name or tag name.  For cherry-picks, we rely on the git
patch-id to find the original commit hash. And for release notes, there
is no need to store it in git notes, because we always manually modify
the generated release note.