bazel syntax: improve error for dynamic use-before-definition
The error emitted for a dynamic use-before-definition in a BUILD file
used to look confusingly like a static error:
  "name x is not defined; did you mean...?"

Such errors are expected in programs like 'print(x); x=1'
where there is no misspelling.

This change makes the error resemble the dynamic error in a .bzl file:
"variable x referenced before assignment", which makes clear that
there is an assignment for x, it just hasn't been executed yet.

PiperOrigin-RevId: 294428556