Fix #10459: Support .bazelversion files without a trailing newline.
This brought up the question of how to handle a completely empty .bazelversion file. I checked what Bazelisk does in this case (silently assume that the latest stable Bazel version should be used, just as if you explicitly write "latest" into the file). Thus we slightly change the behavior of the wrapper to more closely match the behavior of Bazelisk, which will help to reduce user confusion.

RELNOTES: None.
PiperOrigin-RevId: 294438748