Try to deflake MacOSXFsEventsDiffAwarenessTest.
This test was using a short sleep to wait for fsevents to propagate and,
of course, sometimes we exhaust this timeout without seeing the events
(if the machine is very busy, for example).  Or at least that's the
theory behind why this test has become flaky now.

Fix this by retrying the query a few times so that we wait for a very
short amount of time in the common case--but with the ability to tolerate
delays.

While doing this, significantly change how the test works to improve
readability and to cope with the fact that retrying the getDiff calls
requires changing the algorithm we use to check for success (given that
the arguments to this function have to be sequential views).

Tested: Manually ran this test on my Mac with --runs_per_test=100 and
with a very short sleep in each iteration (10ms instead of the 100ms
I'm adding here).

RELNOTES: None.
PiperOrigin-RevId: 294720397