Support always-dirty Ninja actions
Always-dirty phony targets are those which do not have any inputs: "build alias: phony".
Ninja specification says:
"phony can also be used to create dummy targets for files which may not exist at build time. If a phony build statement is written without any dependencies, the target will be considered out of date if it does not exist. Without a phony build statement, Ninja will report an error if the file does not exist and is required by the build."
So actually Ninja leaves the possibility that the phony target file *may* exist, and then it is not always dirty.
However in our case, we in general do not allow explicit expression of the circular dependency on self-outputs (instead, action cache should determine whether the outputs already exist and do not have to be rebuilt), so we will interpret phony actions without inputs as always dirty.

All usual direct dependants of those actions automatically also always-dirty (but not
the transitive dependants: they should check whether their computed inputs have changed).
As phony targets are not performing any actions,

*all phony transitive dependants of always-dirty phony targets are themselves always-dirty.*

That is why we can compute the always-dirty flag for the phony targets, and use it for marking their direct non-phony dependants as actions to be executed unconditionally.