Execute only the subgraph of Ninja targets, needed to produce required outputs
Additionally, we have to modify how we work with phony targets transitive closure of artifacts: we can not compute artifacts for phony targets right away, because artifacts are registered by RuleContext and later appear to be obsolete if we do not create the actions for those artifacts (and since we now executing only a subgraph, we do not create actions for everything).
So we introduce a cache for reusing already computed transitive closures, and we keep information for phony targets which other phony targets they include.
The same cache can be reused for creating the output groups provider.