Declare TestConfiguration as a required fragment in the Java rules.
TestConfiguration is required to check in the Java rules implementation if the persistent test runner is enabled. The check avoids creating new data structures when the persistent test runner is disabled.

Requiring TestConfiguration in the Java rules is unfortunate, but it seems like a safe option. Moving the option --experimental_persistent_test_runner under JavaConfiguration is not possible because the option is used by the TestAction, which has to remain language-agnostic. An alternative is moving the option under CoreOptions, but that isn't accurate and pollutes the current curated set of core options.

PiperOrigin-RevId: 294923678