Fix crashes on interrupt due to concurrent sandboxfs requests.
When we receive an interrupt, we may still have an ongoing request to
sandboxfs to map a sandbox.  In that case, attempting to create a
request to destroy the sandbox will make us crash due to the
sanity-checks we have in RealSandboxfs02Process (and the request will
fail anyway if it reaches sandboxfs).

Prevent this by tracking whether the sandbox needs cleanup and only
issuing the deletion requests if the creation succeeded.

Tested: I've tried to write a test for this but it's very difficult.
To test this condition, we need to send an interrupt while we have a
build with a sandbox creation request in-flight.  But to catch the
latter, we need the requests to be very big.  I've simulated this
behavior by adding a sleep in RealSandboxfs02Process#createSandbox
and killing Bazel during a build.  Would reproduce the crashes
instantly before the fix and haven't been able to after this change.

RELNOTES: None.
PiperOrigin-RevId: 294974301