Handle external mutations to completionFuture in ActionExecutionState. The class assumes that the future will never be mutated except under its synchronized lock, but that is false: https://github.com/bazelbuild/bazel/blob/146d51aa1ec9dcb721a7483479ef0b1ac21d39f1/src/main/java/com/google/devtools/build/lib/concurrent/AbstractQueueVisitor.java#L485 cancels the combined future it's passed if it's interrupted, and more generally any registered futures will be canceled when the pool is interrupted.
In the case that there are three shared actions, this can trigger the following bug: action A starts executing. Actions B and C start executing. The build is interrupted at some point between here and later, but neither B's nor C's threads notice yet. Action B notices action A is already executing and sets completionFuture. It then exits, returning control to AbstractParallelEvaluator$Evaluate#run code. That code tries to register the future with AbstractQueueVisitor. Either here or later, the future is canceled when the pool is interrupted. Action A's execution thread takes some time to notice. After this cancellation, Action C starts running. It too notices Action A is already executing. But then it unexpectedly finds that the future is done.

PiperOrigin-RevId: 295031074