Improve the performance of createFileSystem in WorkerExecRoot
The performance improvement is due to the reduction of the amount
of I/O performed when creating all necessary directories/symlinks.
Previously the code would delete all inputs and then recreate them,
even if some of the symlinks/directories were already pointing to
the right destinations. The new code will avoid deleting/recreating
anything that is already correct according to `SandboxInputs`.

For actions that have 1000-1500 inputs, this brings roughly a 2x
improvement: from about 100-150ms to 40-70ms. This is a non-trivial
amount for persistent workers that already cache a lot in-memory
and can perform some actions in less than 100ms.

RELNOTES: Improve the performance of creating a sandboxed execution root for workers when the number of inputs is large (>1000).
PiperOrigin-RevId: 295529113