Introduce ninja_build for the correct composition of targets.
ninja_graph is responsible now only for parsing the graph and symlinking inputs under output_root. This way it will always be a leaf in the build graph, and all combinations of non-cyclic dependencies between Bazel-built and Ninja-built targets are possible.
ninja_build absorbs parsed NinjaTarget objects through NinjaGraphProvider.
ninja_build creates actions for the subgraph for the targets from output_groups.
Bazel-built dependencies are passed to ninja_build with deps_mapping attribute.
There is an interoperability with Bazel test.

! this CL additionally contains a fix for deps_mapping to also replace the path to bazel-built file in the constructed command line.