Fix --experimental_prefer_mutual_xcode=false to recognize the correct developer directory for local xcode version tools when a mutually available Xcode is selected.
The change in https://github.com/bazelbuild/bazel/commit/2572bba43133692b249eb840d5262cb32bc020db started using the version number of the default local version to find the local xcode, but this results in looking for developer tools at a path that doesn't exist. Instead, use the remote version associated with that local version, and match on that one, as we do with --experimental_prefer_mutual_xcode=true.

RELNOTES: None.
PiperOrigin-RevId: 295961075