Bazel Glossary
First draft of the Bazel Glossary.

There are a few terms left undefined, which I've marked as comments and TODOs for now.

Another follow up is to add out-links to the full documentation page for particular terms, like "Aspect".

Thanks to all contributors to this glossary:

* @gregestren
* @ahumesky
* @laurentlb
* @tetromino
* @jmmv
* @artem-zinnatullin

Fixes https://github.com/bazelbuild/bazel/issues/10541

Closes #10677.

PiperOrigin-RevId: 295967824