Packaging CDS archive file (.jsa) in deploy JAR.
This is the first step of a series of planned changes to support producing a CDS archive at build-time automatically and embedding it into the output deploy JAR as part of the 'java_binary' rule action, when it is opt-in.

This CL augments the singlejar tool to prepend the CDS archive data before the JAR content but after the launcher binary, if '--cds_archive <archive_path>' is specified. The start of the CDS archive within the deploy JAR is aligned at page boundary, which is required by file-backed mapping. The start offset is recorded as an attribute, 'Jsa-Offset' in the deploy JAR META-INF/MANIFEST.MF. The recorded 'Jsa-Offset' is used by the JVM to locate and map the archived data at runtime.

RELNOTES: Add --cds_archive option for embedding CDS archive into deploy JAR.
PiperOrigin-RevId: 295991052