bazel syntax: fix crash in CPU profiler due to initialization race
This change moves the createPipe call into the same thread as startTimer.
Before, it was done in the router thread, so it was possible for PROF
signals to arrive before that thread had gotten started, causing a JVM abort.

This change also re-enables the suppressed test,
whose flakiness was likely from this cause.
Running it 100 times showed zero failures (before: 5).

RELNOTES: N/A
PiperOrigin-RevId: 296931264