C++: Add incompatible flag to deprecate old C++ API.
Libraries to link have been wrapped by the class LinkerInput for a while, however, as we were not sure we would stay with this refactoring, the Starlark API was guarded by the experimental_cc_shared_library.

This refactoring was needed to be able to implement cc_shared_library. Now that we are sure of the design, we are deprecating the old Starlark API and replacing it with the new one using an incompatible flag.

Incompatible issue: #10860

RELNOTES:none
PiperOrigin-RevId: 297109588