Avoid buffering sandbox creation requests in memory.
The previous code constructed the mappings of a sandbox fully in-memory
by collecting all mappings into an array of Mapping objects.  When this
representation was built, the code would then serialize it into
sandboxfs' stream.

This was very wasteful, both in terms of generated garbage (several
objects per action input, which means several thousands of objects per
action), and in terms of consumed CPU time.

To address this, write directly into the stream on which sandboxfs
listens.  This means that the sandbox creation code path is now more
sensitive to performance and there are a couple of things to address
in this area, but I'll get to them later.

RELNOTES: None.
PiperOrigin-RevId: 297187139