Prepare for opensourcing lib/buildtool/*Test.java
Replicated the Google-internal class hierarchy
under //src/test/java/.../lib/buildtool, because
this allows syncing code between the Bazel tree
and the Google-internal tree.

Open-sourced the ActionListenerIntegrationTest
though it's not yet enabled, because Bazel doesn't
support running BuildIntegrationTests yet, because
BazelMockCcSupport doesn't implement the crosstool
setup logic yet.

Created the GoogleBuildIntegrationTestCase class
for Bazel, because this is the superclass of
ActionListenerIntegrationTest and in the
Google-internal code base it is required, and
re-creating an open-source version of it in the
Bazel tree was the easiest approach to
open-sourcing ActionListenerIntegrationTest.

RELNOTES: none
PiperOrigin-RevId: 297353036