Do not create TestResults with null data.
TestRunnerAction attempts to read the test data from a local cache. In case of
a failure to read that, we will create pass a null TestResultData to
TestResult, which is unexpected and results with NullPointerExceptions in the
future.

Add a null check to TestResult constructor for data and prevent the TestResult
from being created in case we fail to read the cached data. This will match the
handling of IOExceptions when creating the cached result, which are logged to
remote storage and prevent the event from being emitted.

PiperOrigin-RevId: 297592861