Fix an initialization race in MacOSXFsEventsDiffAwareness and reenable tests.
MacOSXFsEventsDiffAwareness creates a thread at initialization time to run
the fsevents loop. This process takes two steps: first, registering the
fsevents queue with the kernel, which determines the point in time from
which we want to receive events; and, second, the actual start of the
run loop.

Before returning from init(), we must ensure we have at least completed
the first step of waiting for the queue registration -- or else any
file system operations that happen before the registration are lost.

I thought about splitting the run() method in two parts instead of using
a latch for synchronization but that doesn't work: note that the "run"
part of the loop relies on the current thread's CFRunLoop, which is
different from the one we'd do in the preparatory step outside of the
thread.

Tested: Added a sleep in the thread right before we call run().  Without
this fix, we consistently lose all events from the test, and with this
fix, we get them all.

Fixes #10776.

RELNOTES: None.
PiperOrigin-RevId: 297648494