When a fetch for an Object[] returns, cache the unwrapped contents.
By keeping the cached value as the wrapping ListenableFuture<Object[]>, we have the potential of making multiple fetches for the same fingerprint. This is illustrated by the test case - a fetch for a transitive member will only retain the ListenableFuture until it is resolved, then the cache entry becomes eligible for GC even though the underlying Object[] is still in memory.

I don't know how often this happens, but the change seems harmless and we can use the number of NestedSet reads to see if it makes any difference.

RELNOTES: None.
PiperOrigin-RevId: 298361983