Teach ActionExecutionException and downstream structures about DetailedExitCode
I intend this change to be a pure refactoring. It should entail no
externally observable change in behavior.

This replaces ActionExecutionException's handling of ExitCode values. It
now handles DetailedExitCode values.

The upstream effects include:
* SpawnExecException's translation to ActionExecutionException gives it a
  DetailedExitCode instead of an ExitCode.

SpawnExecException could relay its SpawnResult's FailureDetail, but to
keep this change pure, it does not, yet.

SpawnExecException and ActionExecutionException continue to encode the
idea of a "user error" failure with a null DetailedExitCode value, as
they encoded that idea before with a null ExitCode value. For detailed
user errors this must change, but it isn't changed yet.

The downstream effects include:
* BuildResult handles DetailedExitCode instead.
* BlazeCommandResult now preferably handles DetailedExitCode. It retains
  its ExitCode-only factory methods, for now.
* BuildFailedException handles DetailedExitCode instead.

RELNOTES: None.
PiperOrigin-RevId: 298643766