Use prefix encoding for paths.
sandboxfs now has the ability to track directory paths separately from
each mapping request, essentially encoding paths as an integer that
identifies a directory and the file name within that directory.

This change adds support for this kind of encoding to cut down the
massive size of the reconfiguration messages.  As a result, we use less
CPU time in Bazel and in sandboxfs just dealing with communication
overhead.

RELNOTES: None.
PiperOrigin-RevId: 298895132