Add `canonicalId` to the `Downloader` interface.
When Bazel downloads an external file (via `ctx.download()` or similar,
it supports the concept of a "canonical ID". This ID is used to
disambiguate download requests when the content checksum is unknown and
the URL doesn't change between fetched resources.

Links:
* Design: https://github.com/bazelbuild/proposals/blob/master/designs/2019-04-29-cache.md
* Implementation PR: https://github.com/bazelbuild/bazel/issues/5144
* API doc: https://docs.bazel.build/versions/master/skylark/lib/repository_ctx.html#download

This field was properly plumbed into the `Downloader` interface when it
was added by PR https://github.com/bazelbuild/bazel/pull/10547, but an
ad-hoc change during import caused it to get lost. We need to put it
back, or remote downloaders won't be able to do correct cache lookups
for these resources.