remote/repo: don't accept non-zero exit code from action cache
Some remote execution systems choose to cache action results with a
non-zero exit code. This can be problematic as it is hard to
distinquish an action that fails deterministically from a flaky one.

If a non-zero exit code is returned from cache, then fallback to
re-execute the action.

This change only affects --experimental_repo_remote_exec.