Add sandboxfs support to the processwrapper-sandbox spawn runner.
The lack of this is an overlook from when I first added sandboxfs support
to the linux-sandbox and darwin-sandbox spawn runners.  I noticed this
by chance when doing some sandboxfs work on a Debian VM, where namespaces
are not allowed for unprivileged users.  I saw that my builds were using
the processwrapper-sandbox, not the linux-sandbox, and that they were not
using sandboxfs.

I know that this comes with no tests.  There currently are not tests to
validate a real build with sandboxfs enabled.  If there were, these tests
should verify the behavior with the platform-specific spawn runner and
now also with the more generic processwrapper-sandbox.  However, adding
these tests is pretty complicated and their benefit would be pretty low
at this point (because sandboxfs is not present in Bazel CI so we would
not run them automatically).  I'll work on these later if we decide to
make further progress on this feature.

RELNOTES: None.
PiperOrigin-RevId: 299846783