Ensure the workspace name appears in paths within sandboxfs sandboxes.
Some tools walk path hiearchies looking for the workspace name and
misbehave if that component is not present.  All sandboxed spawn runners
have provisions to deal with this and explicitly add the workspace name
to the execroots they generate... but the sandboxfs spawn runner was not
doing so.  Fix it now.

RELNOTES: None.
PiperOrigin-RevId: 299850812