Cap the number of attributes per rule-class at 150
Intention is to prevent hard-crashes due to runaway RuleClass growth -
AttributeContainer has a 254 cutoff for number of attributes, when that's
hit we crash hard.

150 was chosen to be high enough that it's significantly unlikely to affect
any existing rule classes, and low enough to give us the flexibility to lift
the limit if we absolutely must without running into the AttributeContainer
wall.

RELNOTES: A maximum 150 attributes per RuleClass is enforced
PiperOrigin-RevId: 300348762