Roll forward of https://github.com/bazelbuild/bazel/commit/78fdb7d62d
The original change broke some call paths that didn't have an event reporter to
provide. It's also a bit fragile in relying on AbstractUnorderedFormatter.setEventHandler to provide the reporter - there are different call sites and if we forget to call that method from one of them we'd get a null pointer exception.

This version simply makes the reporting optional: if the reporter is available the output formatter reports to it. Else it skips it.

Ideally we'd do a deeper audit of all call sites and guarantee the existence of a reporter in all expected places. That's a lot of refactoring, and in some disparate pieces of code that serve varying purposes. So it's unclear what assumptions are correct in what place.

This approach, while weaker, at least avoids the crash and provides a clear reason in *some* use cases. We could presumably expand such qualifying use cases incrementally as needed. In the meantime those use cases will simply return misleading data. :(

PiperOrigin-RevId: 300832332