Fix --build_event_publish_all_actions with --remote_download_minimal
The ActionExecutedEvent references the primaryOutput of the executed
action (see ActionExecutedEvent::referencedLocalFiles).
ByteStreamBuildEventArtifactUploader::upload will try to ensure that
this referenced file is present on the remote cache.  It calls
file.isDirectory(), which will eventually go to
RemoteActionFileSystem::getRemoteInputMetadata.  However,
RemoteActionFileSystem::inputArtifactData only contains metadata about
the input files, not action outputs.

This change proposes a fix, where we make RemoteActionFileSystem
implement MetadataInjector, and inject information about output files
from ActionMetadataHandler::injectRemoteFile.

The implementation is likely not correct for all cases, but demonstrates
the issue and a possible approach to fix it.