Add simpler Starlark interface for creating strict_deps
This is in preparation for removing the old interface.  The old
interface requires that user create an ObjcProvider that is then
passed as a "direct_dep_providers" to create another ObjcProvider.
Now that we only support one kind of strict_deps, we don't need the
generality that this interface provides us.  The old interface also
relies on the ObjcProvider compile info API that we are in the process
of obsoleting.

RELNOTES: None
PiperOrigin-RevId: 301247927