Treat timeouts expanding NestedSet as lost inputs.
NestedSetExpander is now permitted to throw TimeoutException, which it may wish to do if the NestedSet is from remote storage and is not available within a certain threshold. In this case, CppCompileAction will treat this dependency's primary output as a lost input, along with any other dependencies with still-pending NestedSets. Rewinding will allow us to rebuild the nodes with the missing NestedSets.

To facilitate setting a custom NestedSetExpander in tests, it is now set in SkyframeActionExecutor#configure instead of passed into the constructor.

PiperOrigin-RevId: 301382884