Merge ByteBufferFragment and ByteFragmentAtOffset and call it FileFragment.
It seemed odd that the object that contains the start and end of a fragment is inside the object that contains the ByteBuffer. It's more logical the other way round, but after going this far, I realized that it's a bit nicer to not have two objects at all.

RELNOTES: None.
PiperOrigin-RevId: 301761088