Delay creation of params files until we know where the writable execroot is.
When using workers or sandboxes, the output tree is not locked until the
spawn completes execution.  However, some virtual inputs, like params files,
have to be materialized before running the spawn.

The previous code was materializing these files directly into the output
tree without previously grabbing the dynamic execution lock, causing trouble.

To fix this, postpone the creation of these files until we know where the
writable execroot for a spawn is (which, when using workers or sandboxes is
in a location separate from the actual execroot).

This fix is gated behind --experimental_delay_virtual_input_materialization
so that we can roll it out separately from a release given the limited testing
we can do on builds that do local execution.  I'll remove the old code paths
once we have flipped the flag.

Problem diagnosed by felly@.

RELNOTES: None.
PiperOrigin-RevId: 301849551