Use DetailedExitCode in AbruptExitException
This allows code that constructs this exception to specify FailureDetail
values. Those are useful for communicating more information about failure
modes than just an exit code.

This modifies BuildTool to consume the exception's DetailedExitCode,
which is the first of many similar uses to be transformed.

---

To keep SkyDoc from depending on protobuf resources, this factors
AbruptExitException (which SkyDoc does not depend on) out of
devtools/build/lib:util (which SkyDoc does depend on). Rules which
previously depended on :util for AbruptExitException now directly
depend on its new rule.

RELNOTES: None.
PiperOrigin-RevId: 302031416