Break TestConstants' dependency on Package FactoryBuilderForTesting
PackageFactoryForTesting can pull in a lot of deps that a lot of things using
TestConstants don't need, so separate it out.

We should also do the same thing for
TestConstants#processSkyframeExecutorForTesting - I'd prefer to do that
separately.

Apologies for the comically named class, it's based on the constant that it was
created from.

RELNOTES: None
PiperOrigin-RevId: 302089555